+++ Site/Middlewares/middleware.js
@@ -1,0 +1,7 @@
+const express = require('express');
+
+const bodyParserMiddleware = express.urlencoded({ extended: true });
+
+module.exports = {
+    bodyParserMiddleware
+};
