+++ Site/Template/login.html
@@ -4,16 +4,16 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="styles.css">
-    <script type="text/javascript" src="../login.js"></script>
+    <script type="text/javascript" src="../Controllers/login.js"></script>
     <title>Page de connexion</title>
 </head>
 <body>
 <div class="login-container">
     <h2>Connexion</h2>
-    <form action="dashboard.html" method="get" name="registration"><!--Changer l'action pour envoyer sur le dashboard-->
+    <form action="/login" method="post" name="registration">
         <label for="email">Adresse email:</label>
         <input type="email" id="email" name="email" required>
-
+    
         <label for="password">Mot de passe:</label>
         <input type="password" id="password" name="password" required>
         <br><br>
